moved forEach and map from store (and cursor) to new helper class Cursors
