+++ lib/src/main/java/com/yheriatovych/reductor/Cursor.java
@@ -15,13 +15,4 @@ public interface Cursor<State> {
      * @return instance of {@link Cancelable} to be used to cancel subscription (remove listener)
      */
     Cancelable subscribe(StateChangeListener<State> listener);
-
-    /**
-     * Creates a Cursor by applying a specified function to state.
-     *
-     * @param mapper the function that will be applied to a state. Must be free of side effects.
-     * @param <T>    the output type
-     * @return Cursor that holds state of type T produced by mapper function
-     */
-    <T> Cursor<T> map(Function<State, T> mapper);
 }
